Adding that image lit af
Since we want all of our assets in the dist folder,
lets add a quick command to copy over our index.html file.
And DONT forget to change your browser url to dist/index.html.
We are creating a single bundle so we want everything in webpack.
Webpack will do this for us later, another option would be to require the
index.html file in the index.js file.