Move cleaning dist code into webpack
It is important to ensure we have a complete bundle in our dist folder
to be shipt after the build. This ensures no stale code is in there and
we can package up our dist folder for production. Now our js code is sort of
a mess now, lets utilize webpack to start oganizing our code a little better.