Adding dev server
Before we start cleaning things up lets make development a little easier.
We will use the webpack-dev-server, which will allow us auto reloading of the
page and we no longer have to rebuild ourselves to check for changes. We have nice source
maps for debugging. You will notice it is much faster, and no longer needs a dist folder since code is
kept in memory.