Break up js code
Now we have the freedom to refactor our code any way we like,
we can move utilities to their own file ect.

Now we have an app that simply defines our app setup code in
the main.js file, then runs the core app code in app.js. This pulls
in the content we need. If we have code that isn't used it will not
be put in the bundle allowing for smaller code. We can go into tree shaking
and other optimizations if we want.